Convert to fp16

+++ jkbc/jkbc/model/architectures/bonito.py
@@ -6,7 +6,7 @@ MODEL_NAME = 'bonito'
 
 def model(window_size, device, base_dir):
     config = toml.load(f"{base_dir}/jkbc/jkbc/model/architectures/quartznet5x5.toml")
-    model = Model(config).to(device=device)
+    model = Model(config).to(device=device).half()
     pred_out_dim = window_size//3+1
     return model, (MODEL_NAME, pred_out_dim)
